% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise.R
\name{CONCISE}
\alias{CONCISE}
\title{CONCISE ligand-receptor interaction analysis}
\usage{
CONCISE(
  pairdb,
  rawcount,
  loc,
  Wx_distance_threshold,
  remove_invalid_genes = TRUE,
  lib_size = NULL
)
}
\description{
CONCISE ligand-receptor interaction analysis
}
