% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise_coexp.R
\name{CONCISE_coexp}
\alias{CONCISE_coexp}
\title{CONCISE spatial gene-gene co-expression analysis}
\usage{
CONCISE_coexp(
  gene_list,
  rawcount,
  loc,
  remove_invalid_genes = TRUE,
  lib_size = NULL
)
}
\description{
CONCISE spatial gene-gene co-expression analysis
}
