% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise_joint_sparse.R
\name{CONCISE_joint_sparse}
\alias{CONCISE_joint_sparse}
\title{CONCISE ligand-receptor interaction analysis: Joint ligand-receptor fitting with sparse matrix for efficient computation}
\usage{
CONCISE_joint_sparse(
  result_path,
  cell_type_anno,
  ligand_cell_type,
  receptor_cell_type,
  pairdb,
  rawcount,
  loc,
  Wx_distance_threshold,
  remove_invalid_genes = TRUE,
  one_side_test = TRUE,
  lib_size = NULL
)
}
\description{
CONCISE ligand-receptor interaction analysis: Joint ligand-receptor fitting with sparse matrix for efficient computation
}
