% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise_marginal_sparse.R
\name{CONCISE_marginal_sparse}
\alias{CONCISE_marginal_sparse}
\title{CONCISE ligand-receptor interaction analysis: Marginal fitting with sparse matrix for efficient computation}
\usage{
CONCISE_marginal_sparse(
  kernel_choice,
  result_path,
  pairdb,
  rawcount,
  loc,
  lib_size = NULL
)
}
\description{
CONCISE ligand-receptor interaction analysis: Marginal fitting with sparse matrix for efficient computation
}
