% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_database.R
\name{InteractionDB}
\alias{InteractionDB}
\title{Organize ligand-receptor interaction database}
\usage{
InteractionDB(pairdb, rawcount, sparsity_threshold)
}
\description{
Organize ligand-receptor interaction database
}
