% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqunifplot.R
\name{qqunif.plot}
\alias{qqunif.plot}
\title{Quantile-quantile plot}
\usage{
qqunif.plot(
  pvalues,
  lim = 5,
  should.thin = T,
  thin.obs.places = 2,
  thin.exp.places = 2,
  xlab = expression(paste("Expected (", -log[10], " p-value)")),
  ylab = expression(paste("Observed (", -log[10], " p-value)")),
  main = "Q-Q plot",
  draw.conf = TRUE,
  conf.points = 1000,
  conf.col = "lightgray",
  conf.alpha = 0.05,
  already.transformed = FALSE,
  pch = 20,
  aspect = "iso",
  prepanel = prepanel.qqunif,
  par.settings = list(),
  ...
)
}
\description{
Quantile-quantile plot
}
